# docs/swagger.yaml
basePath: /v1/users
definitions:
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.DeletePersonalChat:
    properties:
      id:
        type: string
      user_id:
        type: string
    type: object
...
  github_com_Flikest_PingVi_backend_internal_delivery_dto.PersonalChat:
    properties:
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user1_id:
        type: string
      user2_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is PingVi selective forwarding unit.
  termsOfService: http://swagger.io/terms/
  title: PingVi selective forwarding unit
  version: "1.0"
paths:
...
  /personal_chat:
    delete:
      consumes:
      - application/json
      description: |-
        Permanently deletes a personal chat. User must be a participant.
        - Only participants can delete the chat
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Personal chat ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.DeletePersonalChat'
      produces:
      - application/json
      responses:
        "200":
          description: "Chat ID\"\t\t\t\texample({\"chat_id\":\"123e4567-e89b-12d3-a456-426614174000\"})"
          schema:
            additionalProperties: true
            type: object
        "400":
          description: "Invalid request\"\t\texample({\"error\":\"invalid body\"})"
          schema:
            additionalProperties: true
            type: object
        "403":
          description: "Not a participant\"\t\texample({\"error\":\"not belongs to
            chat\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to delete
            personal chat\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a personal chat
      tags:
      - messenger-personal-chat
    post:
      consumes:
      - application/json
      description: |-
        Creates a new direct message chat between two users.
        - If chat already exists, returns existing chat
        - Both users must be authenticated
      parameters:
      - description: Bearer JWT token
        example: '"Bearer eyJhbGciOiJIUzI1NiIs..."'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Personal chat data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.PersonalChat'
      produces:
      - application/json
      responses:
        "201":
          description: Created personal chat
          schema:
            $ref: '#/definitions/github_com_Flikest_PingVi_backend_internal_delivery_dto.PersonalChat'
        "400":
          description: "Invalid request\"\t\texample({\"error\":\"invalid body\"})"
          schema:
            additionalProperties: true
            type: object
        "500":
          description: "Internal server error\"\texample({\"error\":\"failed to create
            personal chat\"})"
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a personal chat
      tags:
      - messenger-personal-chat
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"